backend/requirements.txt
fastapi
uvicorn[standard]
requests
langchain-community
chromadb
python-dotenv
